Session for 31-30-2020
During this session, we worked on implementing scoring.
First, we implemented a function `scoreWord` that takes in
a word and a mapping of letters and points and returns the point
value of the input word. We also laid the foundation for adding score
multipliers, by modifying the parseBoard function. Instead of
returning an array of words (strings), it now returns an array of objects,
that have a `word` key and additional information about its placement on
the board.